+++ src/main/java/com/sprint5team/monew/domain/comment/dto/request/CommentRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.sprint5team.monew.domain.comment;
+package com.sprint5team.monew.domain.comment.dto.request;
 
 import java.util.UUID;
